+++ tests/test_util.py
@@ -1,10 +1,11 @@
 import asyncio
+
 import pytest
 
 from rap.client.processor.crypto import AutoExpireSet
 from rap.client.utils import raise_rap_error
 from rap.common.exceptions import RPCError
-from rap.common.utlis import Event, State, gen_new_param_coro, check_func_type
+from rap.common.utlis import Event, State, check_func_type, gen_new_param_coro
 
 pytestmark = pytest.mark.asyncio
 
@@ -16,7 +17,7 @@ async def demo(a: int, b: int) -> int:
 class TestUtil:
     def test_state_class(self) -> None:
         state: State = State()
-        state.demo = '123'
+        state.demo = "123"
         assert 1 == len(state)
 
         del state.demo
@@ -28,8 +29,8 @@ class TestUtil:
         assert exec_msg.endswith("object has no attribute 'demo'")
 
     def test_event_to_tuple(self) -> None:
-        event: Event = Event('name', 'info')
-        assert ('name', 'info') == event.to_tuple()
+        event: Event = Event("name", "info")
+        assert ("name", "info") == event.to_tuple()
 
     async def test_gen_new_param_coro(self) -> None:
 
@@ -46,19 +47,19 @@ class TestUtil:
         assert exec_msg == "customer_info"
 
     def test_check_func_type(self) -> None:
-
-        def demo(a: int, b: str) -> int: pass
+        def demo(a: int, b: str) -> int:
+            pass
 
         with pytest.raises(TypeError):
             check_func_type(demo, (1, 2))
 
     async def test_auto_expire_set(self) -> None:
         auto_expire_set: AutoExpireSet = AutoExpireSet(1.5)
-        auto_expire_set.add('test1', 0.1)
-        auto_expire_set.add('test2', 1)
-        assert 'test1' in auto_expire_set
+        auto_expire_set.add("test1", 0.1)
+        auto_expire_set.add("test2", 1)
+        assert "test1" in auto_expire_set
         await asyncio.sleep(0.5)
-        assert 'test1' not in auto_expire_set
-        assert 'test2' in auto_expire_set
+        assert "test1" not in auto_expire_set
+        assert "test2" in auto_expire_set
         await asyncio.sleep(1)
-        assert 'test2' not in auto_expire_set
+        assert "test2" not in auto_expire_set